repactor: 로또 가격 클래스 분리
